LP64 fix lpair__Init: long x -> int x prevents zero-extension of negative percentage args
FossilOrigin-Name: 01c6d0e08bdec636fdfd08253da2f0205984a4fb990216b2361645dabe9588a8

+++ src/atk/supportviews/lpair.c
@@ -517,7 +517,7 @@ long	x;
 void lpair__Init(newl, l1, l2, x)
 struct lpair *newl;
 struct view *l1, *l2;
-long	x;
+int	x;
 {
 
 	if (x < 0)
